Add projection test using Geoff's new PD->euler routines
This only works after the realization on how the q2Spider routine
is splitting between Geoff's standard/alternate conventions:

angle = angle_standard if s_x > 0 else angle_alternate

I updated the psi_ang test to reflect this, as well as the code that converts
a set of PDs to a set of Euler angles. Hoozah!